UI-2641: add the feature to sync a number individually, without having to sync the whole list of numbers at once.
The problem we're fixing is that for large accounts (think 3,000+ numbers)
it could take minutes in order to sync all the numbers. Sometimes the
request can crash because of that. And in most cases, users know which
number is out of sync. So we add the feature of "Syncing" an individual
number. You can find this feature in the features dropdown available next
to phone numbers in SmartPBX, Number Manager, PBX Connector, etc...