OOB: Execute insertTemplate() callback without delay
The only restriction on this callback is to call it once the template is
inserted in the DOM. Before, it would wait to be called until the fadeIn
animation ended, adding an unnecessary delay of 250ms in its execution.
Now, it is called right after the insertion of the template, with no
delay.