UI-3116: Move sub-module templates in their own folder (#189)
* Move sub-module templates in their own `views` folder

By using `getTemplate()` we can specify which sub-module a template is
related to, enabling this behavior.
The templates names were shortened by stripping the name of their
sub-module from it since they are stored under their sub-module folder.
String templates are created using `getTemplate()` instead of
`monster.template()` that we are trying to depreciate.

* Use correct folder name for relative path

* Only use `getTemplate()` for both regular and string templates