Loosely check for Kazoo number feature  props
The previous check was strictly enforcing `number` to be a Kazoo phone
number, meaning that objects without the required props would throw an
error.

This behavior is not backward compatible with current usage (PBX
Connector for example) and it would take too long to refacor those use
cases on an app by app basis.

THe solution it to loosely check for the props related to number feature
and return an empty array when missing/already empty.