Update auth token for kazoo connection
Previously, the new auth token was passed down as a request parameter to
be used by appContext#callApi to override the expired token.

This behavior is no reproducible with monster#request as the auth token
is set during request definition so it is not possible to pass it down
as a request parameter.

The solution is simply to use the dedicated setKazooAPIToken method of
the auth app to directly set the auth token, without having to pass down
the token down as request parameter.