Wait for async operations before invoking callback
Loading dependencies is an asynchonous operation since a HTTP request
needs to be made to retrieve the file in question.
Since loading an extension is equivalent to loading an app, it could
potentiallly trigger dependency loading as well, hence making the
operation asynchronous as well.

This problem is solved by simply bundling all dependencies/extensions
loading operations in an array of tasks executed in parallel.

Inifite extensions loading is taken care of by simply checking if an
extension (which is an app) is already loaded locally.