Not generate default error on port state change
The errors are handled within the error callback. If the `callApi`
function is left to handle the errors, it shows a dialog with the error
details in an asynchronous way. This process is stopped abruptly because
the wizard is rendered again, which in turn causes that the error dialog
is not displayed, but the page body's scroll is disabled as if it was
shown.