Enforce app permissions on route switch
It was possible to access an app by its route even when a user did not
have the permissions to use it.

Before loading an app, we now check the relevant permissions
(all/admin/user specific) and when an app is not loadable, we revert to
load the last loaded app if possible, or the default app.