Bubble up missing default i18n file error
We want the app loading process to error out when the default i18n file
is missing on the server as the app cannot load without it.

On the other hand, when a custom language is selected, but the relevant
i18n file is not accessible, the app is still loadable as it will
fallback on the default i18n file, so we don't bubble up an error when
file fetching errors out.